refactor statsEffectEmbed under status effect directory
